list item

+++ app/models.py
@@ -22,4 +22,4 @@ class Item(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     item_name = db.Column(db.String(64), index=True, unique=True)
     item_description = db.Column(db.String(64), index=True, unique=True)
-    item_price = db.Column(db.Integer, index=True, unique=True)
+    item_price = db.Column(db.Float, index=True, unique=True)
